- Updated README - Godbert: Moved AttachedImage into the correct namespace. - SaintCoinach: Changed property names on Graphics.Bone because they were wrong.
